fix: 잔 스타일링 수정
