+++ src/Components/Header/Tab/Hamburger.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import Tab from './Tab'
+
+const Hamburger = () => {
+  return <div>{/* <Tab /> */}</div>
+}
+
+export default Hamburger
